moving reporting to the Evaluator class
